+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 - Changed electrolyzer capacity to be an input to the electrolyzer cost models rather than pulled from the cost model config.
 - Updated finance parameter organization naming in `plant_config`.
 - Added cost model base class and removed `plant_config['finance_parameters']['discount_years']['tech']`. Some cost models require user-input cost year (`tech_config[tech]['model_inputs']['cost_parameters']['cost_year']`) others do not. Cost year is output from cost models as a discrete output.
+- Add ocean alkalinity enhancement technology model.
 - Added ability to export ProFAST object to yaml file in `ProFastComp`
 
 ## 0.3.0 [May 2 2025]